New checks

+++ tidal.py
@@ -21,20 +21,38 @@ def start():
 
 	if int(mode) == 1:
 		playlist_id = input("Enter playlist id: ")
-		playlist = session.get_playlist_tracks(playlist_id=playlist_id)
-		for track in playlist:
-			download_flac(track)
+		try:
+			playlist = session.get_playlist_tracks(playlist_id=playlist_id)
+			can_download = 1
+		except Exception:
+			print("Playlist not found.")
+			can_download = 0
+		if can_download == 1:
+			for track in playlist:
+				download_flac(track)
 		want_start()
 	elif int(mode) == 2:
 		track_id = input("Enter track id: ")
-		track = session._map_request('tracks/'+str(track_id), params={'limit': 100}, ret='tracks')
-		download_flac(track)
+		try:
+			track = session._map_request('tracks/'+str(track_id), params={'limit': 100}, ret='tracks')
+			can_download = 1
+		except Exception:
+			print("Track not found.")
+			can_download = 0
+		if can_download == 1:
+			download_flac(track)
 		want_start()
 	elif int(mode) == 3:
 		album_id = input("Enter album id: ")
-		album = session.get_album_tracks(album_id=album_id)
-		for track in album:
-			download_flac(track)
+		try:
+			album = session.get_album_tracks(album_id=album_id)
+			can_download = 1
+		except Exception:
+			print("Album not found.")
+			can_download = 0
+		if can_download == 1:
+			for track in album:
+				download_flac(track)
 		want_start()
 	elif int(mode) == 4:
 		search_query = input("Enter search query: ")
@@ -56,21 +74,34 @@ def download_flac(track):
 	album_artist = translit(u""+track.album.artist.name, "ru", reversed=True).encode("UTF-8")
 	download(url, 'music/'+name+'.flac')
 	download(track.album.image, 'music/'+name+'.png')
-	audio = FLAC("music/"+name+".flac")
-	albumart = "music/"+name+".png"
-	image = Picture()
-	image.type = 3
-	mime = 'image/png'
-	image.desc = 'front cover'
-	with open(albumart, 'rb') as f: # better than open(albumart, 'rb').read() ?
-		image.data = f.read()
-	audio['artist'] = artist_name
-	audio['title'] = name
-	audio['album'] = album_name
-	audio['date'] = releaseDate
-	audio.add_picture(image)
-	audio.save()
-	os.remove("music/"+name+".png")
+	if os.path.exists("music/"+name+".flac"):
+		flac = 1
+	if not os.path.exists("music/"+name+".flac"):
+		flac = 0
+	if ".flac" not in url:
+		flac = 0
+	if ".flac" in url:
+		flac = 1
+	if flac == 1:
+		audio = FLAC("music/"+name+".flac")
+		albumart = "music/"+name+".png"
+		image = Picture()
+		image.type = 3
+		mime = 'image/png'
+		image.desc = 'front cover'
+		with open(albumart, 'rb') as f: # better than open(albumart, 'rb').read() ?
+			image.data = f.read()
+		audio['artist'] = artist_name
+		audio['title'] = name
+		audio['album'] = album_name
+		audio['date'] = releaseDate
+		audio.add_picture(image)
+		audio.save()
+		os.remove("music/"+name+".png")
+	elif flac == 0:
+		print(name+" isn't in flac or can't be downloaded in this country.")
+	else:
+		print("Unknown error.")
 
 def download(url, file_name):
 	with open(file_name, "wb") as file:
